Revision all id int to long

+++ src/main/java/com/springboot/moa/result/model/PostDetailResultRes.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostDetailResultRes {
-    int resultDetailId;
+    long resultDetailId;
 }
